add label_binarizer in preprocess step
