+++ README.md
@@ -70,10 +70,12 @@ your own data under development.
 ```
 Usage: grants_tagger preprocess wellcome-science [OPTIONS] [INPUT_PATH]
                                                  [OUTPUT_PATH]
+                                                 [LABEL_BINARIZER_PATH]
 
 Arguments:
   [INPUT_PATH]   path to raw Excel file with tagged or untagged grant data
   [OUTPUT_PATH]  path to JSONL output file that will be generated
+  [LABEL_BINARIZER_PATH] path to pickle file that will contain the label binarizer
 
 Options:
   --text-cols TEXT  comma delimited column names to concatenate to text
@@ -86,10 +88,12 @@ Options:
 ```
 Usage: grants_tagger preprocess bioasq-mesh [OPTIONS] [INPUT_PATH]
                                             [OUTPUT_PATH]
+                                            [LABEL_BINARIZER_PATH]
 
 Arguments:
   [INPUT_PATH]   path to BioASQ JSON data
   [OUTPUT_PATH]  path to output JSONL data
+  [LABEL_BINARIZER_PATH] path to pickle file that will contain the label binarizer
 
 Options:
   --mesh-tags-path TEXT      path to mesh tags to filter
